feat: Add fuzzy search functionality using fuse.js
