+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "@local/flash-find",
-  "version": "1.0.3",
+  "version": "1.1.0",
   "description": "blazing fast client-side search utility",
-  "main": "dist/648.bundle.js",
+  "main": "dist/bundle.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "webpack"
